refactor(ai): align activity ownership with message boundaries
