+++ infrastructure/ai/types.ts
@@ -1,6 +1,9 @@
 // AI Provider types
 import defaultCommandBlocklist from '../../lib/commandBlocklist.json';
+import type { AgentActivity, AgentUsage } from '../../domain/agentActivity';
 import type { ProviderContinuation } from './providerContinuation';
+
+export type { AgentActivity, AgentUsage } from '../../domain/agentActivity';
 
 export type AIProviderId =
   | 'openai'
@@ -111,44 +114,6 @@ export interface AIDraft {
 export type AIPanelView =
   | { mode: 'draft' }
   | { mode: 'session'; sessionId: string };
-
-export type AgentActivityStatus = 'running' | 'completed' | 'failed';
-export type AgentFileChangeKind = 'add' | 'delete' | 'update';
-
-export type AgentActivity =
-  | {
-      id: string;
-      type: 'file_change';
-      status: Exclude<AgentActivityStatus, 'running'>;
-      changes: Array<{ path: string; kind: AgentFileChangeKind }>;
-    }
-  | {
-      id: string;
-      type: 'web_search';
-      status: Exclude<AgentActivityStatus, 'failed'>;
-      query: string;
-    }
-  | {
-      id: string;
-      type: 'plan_update';
-      status: Exclude<AgentActivityStatus, 'failed'>;
-      items: Array<{ text: string; completed: boolean }>;
-    }
-  | {
-      id: string;
-      type: 'warning';
-      status: 'completed';
-      message: string;
-    };
-
-export interface AgentUsage {
-  inputTokens: number;
-  cachedInputTokens?: number;
-  outputTokens: number;
-  reasoningTokens?: number;
-  totalTokens: number;
-  estimated?: boolean;
-}
 
 export interface ChatMessage {
   id: string;
